09.19
스킬 인벤 <-> 플레이어 스킬인벤 드래그 앤 드롭 상호작용 완
비활성화된 아이콘 툴팁도 재대로 나오게 했음
스킬 습득안했을 시 아이콘 비활성화
스킬 습득 버튼 생성

스킬 인벤토리 UI 수정

인벤 정보 TMP_Text로 변경 완료
오브젝트의 스탯, HP,MP 등 float => int로 데이터자료 변환

스킬 레벨업 이벤트 진행중

+++ Assets/02.Scripts/UI/PlayerUIManager.cs
@@ -33,10 +33,6 @@ public class PlayerUIManager : MonoBehaviour
     [SerializeField] private Text AttackSpeedText;
     [SerializeField] private Text AttackRangeText;
 
-    //[Header("Skill Info")]
-    //[SerializeField] private Image[] skill_Image           = new Image[6];
-    //[SerializeField] private Image[] skill_CoolTimeImage   = new Image[6];
-    //[SerializeField] private TMP_Text[] skill_CoolTimeText = new TMP_Text[6];
 
     private PlayerController playerCtr;
 
@@ -49,17 +45,13 @@ public class PlayerUIManager : MonoBehaviour
         instance = this;
         DontDestroyOnLoad(this.gameObject);
 
-        //for (int i = 0; i < skill_Image.Length; i++)
-        //{
-        //    skill_CoolTimeText[i] = skill_CoolTimeImage[i].GetComponentInChildren<TMP_Text>();
-        //}
     }
     void Start()
     {
         playerCtr = PlayerController.instance;
     }
 
-    public void SetHPbar(float _playerHP, float _playerMaxHP)
+    public void SetHPbar(int _playerHP, int _playerMaxHP)
     {
         float amount = (_playerHP / _playerMaxHP);
         hpBarImage.fillAmount = amount;
@@ -73,7 +65,7 @@ public class PlayerUIManager : MonoBehaviour
         hpText.text = $"{_playerHP}/{_playerMaxHP}";
     }
 
-    public void SetMPbar(float _playerMP, float _playerMaxMP)
+    public void SetMPbar(int _playerMP, int _playerMaxMP)
     {
         float amount = (_playerMP / _playerMaxMP);
         mpBarImage[0].fillAmount = amount;
@@ -100,59 +92,6 @@ public class PlayerUIManager : MonoBehaviour
         expHandleRT.anchorMax = vec;
         expText.text = $"{_playerEXP}/{_playerMaxEXP} XP";
     }
-
-    //public IEnumerator SkillUsedWarring(int _idx)
-    //{
-    //    WaitForSeconds waitTime = new WaitForSeconds(0.05f);
-
-    //    for (int i = 0; i < 4; i++)
-    //    {
-    //        if (i % 2 == 0) skill_Image[_idx].color = Color.red;
-    //        else skill_Image[_idx].color = Color.white;
-    //        yield return waitTime;
-    //    }
-
-    //    yield break;
-    //}
-
-    //public IEnumerator StartSkillCoolTime(int _idx, float _coolTime, SkillData _curSkill)
-    //{
-    //    _curSkill.SetInUse(true); // 현재 스킬은 사용 중으로 bool 값 상태
-
-    //    skill_CoolTimeImage[_idx].gameObject.SetActive(true);
-    //    skill_CoolTimeImage[_idx].fillAmount = 1;
-    //    float fcoolTimeText = _coolTime;
-    //    float t = 0.0f;
-
-    //    while (0.0f < fcoolTimeText)
-    //    {
-    //        fcoolTimeText -= Time.deltaTime;
-    //        t = Time.deltaTime / _coolTime;
-    //        skill_CoolTimeImage[_idx].fillAmount -= t;
-
-    //        skill_CoolTimeText[_idx].text = $"{(int)fcoolTimeText}";
-    //        yield return null;
-    //    }
-    //    //쿨타임이 끝나면 Fill 이미지와 CoolTime Text 비활성화
-    //    //스킬 사용 가능 상태로 변환
-    //    skill_CoolTimeImage[_idx].gameObject.SetActive(false);
-    //    _curSkill.SetInUse(false);
-    //}
-
-
-
-    //public void UpdateSkill_Image(SkillData[] _skill_datas)
-    //{
-    //    //배운 스킬의 개수만큼 스킬 이미지 활성화
-    //    int size = _skill_datas.Length;
-    //    for (int i = 0; i < size; i++)
-    //    {
-    //        if (_skill_datas[i] == null) continue;
-    //        skill_Image[i].sprite = _skill_datas[i].GetSkill_Sprite();
-    //        skill_Image[i].gameObject.SetActive(true);
-    //    }
-
-    //}
 
     public void DisplayInfo(int _level, float _maxHP, float _maxMP, float _str)
     {
